+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Pink palettes and swatches - [#53](https://github.com/dartoos-dev/eo_color/issues/53)
+  - **Palettes:** Pink, PinkAccent.
+  - **Swatches:** Pinks, PinkAccents.
 
 ## [0.0.9] - 2021-05-21
 ### Added